ImageTool: Fixed code style.

+++ BaseTools/ImageTool/ImageTool.h
@@ -108,7 +108,7 @@ typedef struct {
 	char                                Name[];
 } DebugData;
 
-#define PAGE(x)     ((x) & ~(uint64_t) 4095ULL)
+#define PAGE(x)     ((x) & ~4095U)
 #define PAGE_OFF(x) ((x) & 4095U)
 
 typedef struct {
